# Refacto geojson parsing and converting features