--- a/test/examples/wfs.js
+++ b/test/examples/wfs.js
@@ -12,7 +12,8 @@ describe('wfs', function _() {
 
     it('should pick the correct building', async () => {
         // test picking
-        const buildingId = await page.evaluate(() => picking({ x: 342, y: 243 }).properties.id);
-        assert.equal(buildingId, 'bati_indifferencie.5751442');
+        const buildingId = await page.evaluate(() =>
+             picking({ x: 342, y: 243 }));
+        assert.equal(buildingId.id, 'bati_indifferencie.5751442');
     });
 });