## Description
**Refacto `GeojsonParser,` `Feature2Mesh` and `feature2texture`** 

- add merge features option (so the convert creates only one mesh) 
- Use one array of number to store the coordinates

**BREAKING CHANGE**
One array of number to store the coordinates instead of array of coordinates

## Motivation and Context
- **fix #731 bad performance**

   - Before bug from PR #691, total time by functions are :
     - ` geojsonParser.parse`  320 ms
     - `feature2Mesh`  216 ms
         `parse + convert` 536 ms
     - `renderScene` 350 ms
     - **total** 866 ms

   - From introduction of bug with PR #691, total time by functions are :
     - ` geojsonParser.parse`  874 ms (x 2.7 slower)
     - `feature2Mesh`  451 ms (x 2.1 slower)
        `parse + convert ` 1325 ms (x 2.5 slower)
     - `renderScene` 885 ms (x 2.5 slower)
     - **total** 2210 ms (x 2.5 slower)

   - After fix from this PR, total time by functions are :
     - ` geojsonParser.parse`  90 ms (x 10 faster)
     - `feature2Mesh`  115 ms (x 4 faster)
        `parse + convert ` 205 ms (x 6.5 faster)
     - `renderScene` 314 ms (x 2.8 faster)
     - **total** 724 ms (x 3 faster)

- **ref: Improving the Feature type #763**