@gliegard 2018-10-12
Great performance fix !!