@gchoqueux 2018-10-12
i must fix GUI pixel and line size