@gchoqueux 2018-10-15
Branch is rebased